+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2020-02-02
+
+### Fixes
+- Fix getRedirectionURL hook type
+- No redirectToPath in email verification screen
+
+## [0.5.2] - 2020-02-01
+
+### Changed
+- Redirect to intended page post authentication
 
 ## [0.5.1] - 2020-01-27